feature: Add helpers for date format & total time
Adapter and Controller were setup but are not yet
utilized and may be removed in later commit. Added
acceptance tests for report and time entry links as
well as for report content. Will need to be modified
when CRUD operations are added. All pages get time-entry
model from application route now. This route uses a promise
to load content from JSON file.